which_opts/python/requirements.txt
click==8.1.3
python-magic==0.4.27
